--HG--
extra : commitid : 7z190Fq6FOA

+++ testing/mozharness/scripts/b2g_build.py
@@ -147,7 +147,7 @@ class B2GBuild(LocalesMixin, PurgeMixin,
             'balrog_credentials_file': 'oauth.txt',
             'build_resources_path': '%(abs_obj_dir)s/.mozbuild/build_resources.json',
             'virtualenv_modules': [
-                'requests==2.2.1',
+                'requests==2.8.1',
             ],
             'virtualenv_path': 'venv',
         }
@@ -1147,7 +1147,8 @@ class B2GBuild(LocalesMixin, PurgeMixin,
         self.info("Cleanup .userconfig file.")
         dirs = self.query_abs_dirs()
         userconfig_path = os.path.join(dirs["work_dir"], ".userconfig")
-        os.remove(userconfig_path)
+        if os.path.exists(userconfig_path):
+            os.remove(userconfig_path)
 
 # main {{{1
 if __name__ == '__main__':
